requirements.txt
Django>=3.0,<=3.1
whitenoise
psycopg2-binary
django-auth-adfs
django-storages[azure]
